// src/generated-api/gameserver/models/gameserver-game-session-dto.ts
// tslint:disable
/**
 * GameServer api
 * Matches, players, mmrs
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { GameserverMatchInfoDto } from './gameserver-match-info-dto';

/**
 * 
 * @export
 * @interface GameserverGameSessionDto
 */
export interface GameserverGameSessionDto {
    /**
     * 
     * @type {string}
     * @memberof GameserverGameSessionDto
     */
    url: string;
    /**
     * 
     * @type {number}
     * @memberof GameserverGameSessionDto
     */
    matchId: number;
    /**
     * 
     * @type {GameserverMatchInfoDto}
     * @memberof GameserverGameSessionDto
     */
    info: GameserverMatchInfoDto;
}